Add configure in Lua
This is the first attempt to write the configurator in Lua.

The most interesting thing is that, as long as the generated
configuration file respects the format defined for the installer,
the configurator can be made in any language or technology and
everything will work fine. Just run egginstall and as long as it
knows the location of this configuration file, everything will work fine.